+++ example/react_icu_withHOC/src/i18n.js
@@ -1,0 +1,31 @@
+import i18n from 'i18next';
+import Backend from 'i18next-xhr-backend';
+import LanguageDetector from 'i18next-browser-languagedetector';
+import { reactI18nextModule } from 'react-i18next';
+import ICU from 'i18next-icu'
+
+i18n
+  .use(ICU)
+  .use(Backend)
+  .use(LanguageDetector)
+  .use(reactI18nextModule)
+  .init({
+    fallbackLng: 'en',
+
+    // have a common namespace used around the full app
+    ns: ['translations'],
+    defaultNS: 'translations',
+
+    debug: true,
+
+    interpolation: {
+      escapeValue: false, // not needed for react!!
+    },
+
+    react: {
+      wait: true
+    }
+  });
+
+
+export default i18n;