+++ src/Trans.js
@@ -99,7 +99,7 @@ export default class Trans extends React.Component {
 
   render() {
     const contextAndProps = { i18n: this.context.i18n, t: this.context.t, ...this.props };
-    const { children, count, parent, i18nKey, ns: namespace, i18n, t: tFromContextAndProps, ...additionalProps } = contextAndProps;
+    const { children, count, parent, i18nKey, tOptions, ns: namespace, i18n, t: tFromContextAndProps, ...additionalProps } = contextAndProps;
     const t = tFromContextAndProps || i18n.t.bind(i18n);
 
     const reactI18nextOptions = (i18n.options && i18n.options.react) || {};
@@ -108,7 +108,7 @@ export default class Trans extends React.Component {
     const defaultValue = nodesToString('', children, 0);
     const hashTransKey = reactI18nextOptions.hashTransKey;
     const key = i18nKey || (hashTransKey ? hashTransKey(defaultValue) : defaultValue);
-    const translation = key ? t(key, { interpolation: { prefix: '#$?', suffix: '?$#' }, defaultValue, count, ns: namespace }) : defaultValue;
+    const translation = key ? t(key, { ...tOptions, ...{ interpolation: { prefix: '#$?', suffix: '?$#' }, defaultValue, count, ns: namespace } }) : defaultValue;
 
     if (reactI18nextOptions.exposeNamespace) {
       let ns = typeof t.ns === 'string' ? t.ns : t.ns[0];