+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### upcoming
+- removes componentWillMount lifecycle method from I18n.js render props as it will get deprecated on react 16.3 [404](https://github.com/i18next/react-i18next/issues/404)
+
 ### 7.5.1
 - Fix reference to props in constructor [PR411](https://github.com/i18next/react-i18next/pull/411)